+++ novel_downloader/core/downloaders/__init__.py
@@ -11,9 +11,9 @@ Each downloader is responsible for orchestrating the full lifecycle
 of retrieving, parsing, and saving novel content for a given source.
 """
 
-from .common_asynb_downloader import CommonAsyncDownloader
-from .common_downloader import CommonDownloader
-from .qidian_downloader import QidianDownloader
+from .common import CommonDownloader
+from .common_async import CommonAsyncDownloader
+from .qidian import QidianDownloader
 
 __all__ = [
     "CommonAsyncDownloader",
