+++ novel_downloader/core/downloaders/common.py
@@ -21,7 +21,7 @@ from novel_downloader.utils.file_utils import save_as_json, save_as_txt
 from novel_downloader.utils.network import download_image_as_bytes
 from novel_downloader.utils.time_utils import calculate_time_difference
 
-from .base_downloader import BaseDownloader
+from .base import BaseDownloader
 
 logger = logging.getLogger(__name__)
