+++ novel_downloader/core/factory/__init__.py
@@ -8,18 +8,18 @@ This package provides factory methods for dynamically retrieving components
 based on runtime parameters such as site name or content type.
 """
 
-from .downloader_factory import (
+from .downloader import (
     get_async_downloader,
     get_downloader,
     get_sync_downloader,
 )
-from .parser_factory import get_parser
-from .requester_factory import (
+from .parser import get_parser
+from .requester import (
     get_async_requester,
     get_requester,
     get_sync_requester,
 )
-from .saver_factory import get_saver
+from .saver import get_saver
 
 __all__ = [
     "get_async_downloader",
