+++ novel_downloader/core/parsers/common/helper.py
