+++ novel_downloader/core/parsers/qidian/browser/chapter_normal.py
