+++ novel_downloader/core/parsers/qidian/session/chapter_encrypted.py
