+++ novel_downloader/core/parsers/qidian/shared/helpers.py
