+++ novel_downloader/core/savers/common/common_txt.py
